Revert "Spring quartz"
