+++ back/src/main/java/com/buying/back/application/account/repository/AccountRepository.java
@@ -14,5 +14,10 @@ public interface AccountRepository extends JpaRepository<Account, Long>,
 
   Optional<Account> findByEmail(String email);
 
+  /*
+
+ @Query(value = "select u.* from user u where BINARY(u.login_id) = :loginId", nativeQuery = true)
+  Optional<User> findByBinaryId(@Param("loginId") String loginId);
+   */
 
 }
