+++ back/src/main/java/com/buying/back/infra/config/webconfig/WebConfig.java
@@ -3,9 +3,7 @@ package com.buying.back.infra.config.webconfig;
 import com.buying.back.infra.config.webconfig.inquiry.converter.InquiryChildTypeConverter;
 import com.buying.back.infra.config.webconfig.inquiry.converter.InquiryParentTypeConverter;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Profile;
 import org.springframework.format.FormatterRegistry;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 @Configuration
@@ -16,17 +14,5 @@ public class WebConfig implements WebMvcConfigurer {
     WebMvcConfigurer.super.addFormatters(registry);
     registry.addConverter(new InquiryParentTypeConverter());
     registry.addConverter(new InquiryChildTypeConverter());
-  }
-
-  @Profile("local")
-  @Override
-  public void addCorsMappings(CorsRegistry registry) {
-    System.out.println("cors");
-    registry.addMapping("/**")
-      .allowedOrigins("*")
-      .allowedOriginPatterns("*")
-      .allowedMethods("GET", "POST", "PUT", "DELETE")
-      .allowedHeaders("*")
-      .allowCredentials(true);
   }
 }
